Splitting OSOCE tests into multiple namespaces
