+++ test/Lombiq.OSOCE.Tests.UI/Tests/ModuleTests/BehaviorApplicationInsightsTests.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Lombiq.OSOCE.Tests.UI.Tests;
+namespace Lombiq.OSOCE.Tests.UI.Tests.ModuleTests;
 
 public class BehaviorApplicationInsightsTests : UITestBase
 {
