+++ test/Lombiq.OSOCE.Tests.UI/Tests/ModuleTests/BehaviorHelpfulExtensionsTests.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Lombiq.OSOCE.Tests.UI.Tests;
+namespace Lombiq.OSOCE.Tests.UI.Tests.ModuleTests;
 
 public class BehaviorHelpfulExtensionsTests : UITestBase
 {
