+++ test/Lombiq.OSOCE.Tests.UI/Tests/RetrievalDataTablesTests.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Lombiq.OSOCE.Tests.UI.Tests;
+namespace Lombiq.OSOCE.Tests.UI.Tests.ModuleTests;
 
 public class BehaviorDataTablesTests : UITestBase
 {
