+++ test/Lombiq.OSOCE.Tests.UI/Tests/ThemeTests/BehaviorBaseThemeTests.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Lombiq.OSOCE.Tests.UI.Tests;
+namespace Lombiq.OSOCE.Tests.UI.Tests.ThemeTests;
 
 public class BehaviorBaseThemeTests : UITestBase
 {
